Creating team and generate member without token
